menu permission (emp complete, pos running)

+++ resources/views/suppliers/index.blade.php
@@ -52,8 +52,10 @@ Supplier List
                           <th>Supplier Name</th>
                           <th>Mobile Number</th>
                           {{-- <th>Official Address</th> --}}
-                          <th>Active Status</th>                         
+                          <th>Active Status</th>
+                          @if( (auth()->user()->role_id == 1) || (auth()->user()->role_id == 2))                         
                           <th>Action</th>
+                          @endif
                         </tr>
                         </thead>
                         <tbody>
@@ -72,11 +74,12 @@ Supplier List
                            @endif
                          </td>
                          
+                         @if( (auth()->user()->role_id == 1) || (auth()->user()->role_id == 2))
                           <td>
                             <a href="{{route('edit_supplier',$supplier->id)}}" style="color: white"><button class="btn btn-outline-primary"><i class="fa-solid fa-pen-to-square"></i> Edit</button></a>
                             <button class="btn btn-outline-danger" onclick="deleteOperation({{$supplier->id}})"><i class="fa-solid fa-trash"></i> Delete</button>
                           </td>
-                        
+                        @endif
                         </tr> 
                         @endforeach
